+++ addons/l10n_hr_account_base/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from . import res_partner
+from . import res_company
+from . import res_partner_bank
+from . import account_move
+from . import account_journal